Split awaiting and participant with plans typestate into two files
